+++ sami2py/tests/test_utils.py
@@ -172,3 +172,33 @@ class TestFourierFunction():
         y = sami2py.utils.return_fourier(self.x, self.coeffs)
         target = np.sin(np.pi * self.x / 12.)
         assert (y == target).all()
+
+
+class TestFourierFit():
+    """Test basic functionality of the fourier fitting function"""
+    def setup(self):
+        """Setup the sin wave to be fit by the fitting routine"""
+        self.lt = np.linspace(0, 24, 49)
+        self.coeffs = np.zeros((10, 2))
+        self.coeffs[0, 0] = 1
+        self.v = sami2py.utils.return_fourier(self.lt, self.coeffs)
+
+    def teardown(self):
+        """Delete the setup parameters"""
+        del self.lt, self.coeffs, self.v
+
+    def test_fit(self):
+        """Test the goodness of fit"""
+        v0, fit_coefs, cov = sami2py.utils.fourier_fit(self.lt, self.v, 10)
+        max_diff = np.max(np.abs(self.coeffs.flatten() - fit_coefs.flatten()))
+        assert max_diff < .0000001
+        assert v0 < .00000001
+
+    def test_warning(self):
+        """Test that the warning is generated properly"""
+        nan_drifts = np.array([np.nan])
+        with pytest.warns(Warning):
+            v0, fit_coefs, cov = sami2py.utils.fourier_fit(self.lt, nan_drifts, 10)
+            assert v0 == 0
+            assert (fit_coefs == np.zeros((10, 2))).all()
+            assert (cov == np.zeros((10, 2))).all()